// backend/src/types.ts
export type Identifier = string
// ...
export enum EntityTypes {
  object = 'object',
  layer = 'layer',
  animation = 'animation',
}
// ...
export type Animate = {
  type: 'animate'
  animationIdentifier: Identifier
  entity: EntityTypes.layer | EntityTypes.object
  entityIdentifier: Identifier
  action: Action
}

export enum Action {
  start = 'start',
  stop = 'stop',
  speedUp = 'speedUp',
  slowDown = 'slowDown',
  reset = 'reset',
}